Fixes StaticDisambiguationUnit where an enum was mistakenly private.
The StaticVariant enum is returned and thus used by another parsing
unit. Therefore the enum should indeed be public. It is placed alone and
preceding the private members since this is a convention in this
codebase since these enums would, in general, require being forward
referenced by private methods and members.