Fixes bug in Lexer where an Integer followed by two dots wasn't being tokenized correctly.
Since something like 5..6 would enter the FloatingPointLiteral state of
the lexer, it would register as 5. and then yield an error. Instead, I
added a single character look ahead that would not transfer to
FloatingPointLiteral, but instead return the 5 as an Integer token and
then parse the Slice token on the next run.